requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#

dj-database-url==0.4.1
django-custom-user==0.6
django-graphiql==0.4.4
django-webpack-loader==0.3.0
Django==1.9.7             # via django-custom-user, django-graphiql, graphql-django-view
graphene==0.10.2
graphql-core==0.5.3       # via graphene, graphql-django-view, graphql-relay
graphql-django-view==1.3
graphql-relay==0.4.4      # via graphene
iso8601==0.1.11           # via graphene
promise==0.4.2            # via graphql-core, graphql-relay
psycopg2==2.6.1
py==1.4.31                # via pytest
pytest-django==2.9.1
pytest==2.9.2             # via pytest-django
six==1.10.0               # via graphene, graphql-core
